Attempt at improving detection of localized strings
* Avoid exceptions and overflowing call stack size when matching localized strings
* Introduce Result construct to help us manage results
* No longer returning flat arrays of translation keys from semantic actions - instead, keeping a parse tree of Result objects
* Allow semantic actions to decide whether something is a translation key or not, or whether its parse tree contains references to other localized strings. Consider this example:

tr("this is a string with {token}", {"token": tr("my localized string")}) vs tr("this is a string with {token}", {"token": "my localized string"}).

In the first case - there are two localized strings; in the second - only one.

tr("this is a string with {token}", "a description", {"token": tr("my localized string")}) vs tr("this is a string with {token}", "a description", {"token": "my localized string"})

Difference being - first string args are used to construct a key with label and description...

It should be possible to also improve detection of statements like these:

tr((something ? "fist string": "second string"))

... and by improving I mean actually writing out two localized strings.

This isn't too complete yet, as the only semantics that were updated are for js and objc; and objc is a bit limp with compound statements.