Make token iteration friendlier with subclasses
instead of keeping a local variable with current index, which we'd increment, use a simple iterator, and wrap tokens array into its own struct. This way, we can pass TokenArray instance around, and the associated iterator will be kept in sync. This way, subclasses can iterate over the tokens as they please, and superclass can break up the logic