Attempt to address markup languages
those are the ones that utilize their own markup within HTML/XML - which is basically every JS framework that allows templates. Starting with backbone - as this markup seems to be more widely used (<% %> and <%= %>)

Idea here is that we treat input as XML, we then distinguish strings (either attribute values of tag inner content) as either strings or interpolated strings. Each markup language declares rules for string interpolation. When parsing out interpolated strings, we grab the interpolation content and use a generic (TokenParser for 'base' language) to parse out contents...  This can be overridden by subclasses, utilizing specific token parser...